fixed rubocop
