+++ lib/devise_jwt_controllers.rb
@@ -2,6 +2,7 @@ require 'devise/jwt/controllers/version'
 
 require 'devise/jwt/controllers/engine'
 require 'devise/jwt/controllers/railtie'
+require_relative 'devise/jwt/extensions/warden_config'
 
 require_relative 'devise/jwt/extensions/action_dispatch'
 require_relative 'devise/jwt/overrides/devise'
